# Lancamentos.Api.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lancamentos.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/lancamentos/{id}": {
      "put": {
        "tags": [
          "Lancamentos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLancamentoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLancamentoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLancamentoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Lancamento"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lancamento"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lancamento"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "get": {
        "tags": [
          "Lancamentos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Lancamento"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lancamento"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lancamento"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Lancamentos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/lancamentos": {
      "get": {
        "tags": [
          "Lancamentos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lancamento"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lancamento"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lancamento"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Lancamentos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLancamentoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLancamentoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLancamentoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateLancamentoRequest": {
        "type": "object",
        "properties": {
          "tipoLancamentoId": {
            "type": "string",
            "format": "uuid"
          },
          "descricao": {
            "type": "string",
            "nullable": true
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Lancamento": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tipoLancamentoId": {
            "type": "string",
            "format": "uuid"
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "descricao": {
            "type": "string",
            "nullable": true
          },
          "tipoLancamento": {
            "$ref": "#/components/schemas/TipoLancamento"
          }
        },
        "additionalProperties": false
      },
      "TipoLancamento": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tipo": {
            "type": "string",
            "nullable": true
          },
          "lancamentos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lancamento"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateLancamentoRequest": {
        "type": "object",
        "properties": {
          "valor": {
            "type": "number",
            "format": "double"
          },
          "descricao": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}